requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
jinja2==3.1.4
bcrypt==4.2.0
reportlab==4.2.4
aiofiles==24.1.0
itsdangerous==2.2.0
pywebview==5.3.2
pyinstaller==6.11.1
